Add LocalCheck handler
It's sort of a PoC implementation so far, just to get an idea of how to
work with Kite and its handlers. Still not sure if I've mocked it the
right way, though.

Also, currently all checks are being executed one after another, even
though it does make sense to run them in parallel.