Log each request and response
I couldn't find a way to execute post-handler if an error occurs, so now
the handler should know there's a post-handler for it and it has to be
called manually.

Also, wasn't able to get the result from a handler in a post-handler in
order to avoid manually writing it to request's context.